Altered URL process and fixed syntax, added WIP step icons
Base URL should be written without a trailing forward slash, and extensions will lead with a forward slash. URL is fully assembled prior to being sent to the TruCapGet and TruCapDelete functions.
Step icons are implemented, but they should be replaced by something that fits in a square more comfortably.